HI/requirements.txt
flet>=0.80.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.0.0
